Compute SLAM in intermediary BASE coordinate system
Instead of computing motion and maps of LiDAR sensor in World coordinates,
we can specify an intermediary BASE coordinate system on the moving system.
To sum up, three 3D coordinates system are used :
 - LIDAR {L} : attached to the geometric center of the LiDAR sensor. The
   coordinates of the received pointclouds are expressed in this system.
   LIDAR is rigidly linked (static transform) to BASE.
 - BASE  {B} : attached to the origin of the moving body (e.g. vehicle). We
   are generally interested in tracking an other point of the moving body than
   the LiDAR's (for example, we prefer to track the GPS antenna pose).
 - WORLD {W} : The world coordinate system {W} coincides with BASE at the
   initial position. The output trajectory describes BASE origin in WORLD.

Keypoints extraction is done in a separate function, then transformation is performed,
and at last keypoints bounds are computed.
Keypoints bounds computation is removed from SpinningSensorKeypointExtractor.

This also fixes a little bug on SLAM output World Transform.
The frame_id of this pose was not set correctly to World but to Lidar.