[PV] Hide Ego-Motion parameters if unused, update parameters names and doc
If Ego-motion registration is disabled, many parameters are unnecessary.
In this case, we hide them from GUI.

As Mapping registration paramters are always necessary and will always be there,
this rename them to remove the 'M' suffix at the end, and add 'EM' suffix for
Ego-Motion parameters.

Update parameters doc when several lines are necessary to properly skip lines.