[PV] Disable display mode by default
Display mode is an advanced mode that allows returning additional info to debug or understand SLAM behavior.
As it can be slow to get and copy all this additional data to Poly Data arrays, it should be disabled by default.

This commit also moves DisplayMode parameter to the "filter outputs options" property group.