Rename Mapping step to Localization
This word of "Mapping" was poorly chosen, as it is not mapping, but estimating current pose using previous map.
"Localization" word seems better, as it is the one used in the name SLAM : Simultaneous Localization And Mapping.
Now, it is clearer to see which step does what, and we avoid confusion.