Fix install paths
This project is composed of 1 shared lib LidarSlam and its header files.
We just need one installation path to know where to install it, not 2.

Additionnaly, if the Paraview plugin is enabled, an other path an be set
to install LidarSlamPlugin to user-defined directory (e.g. the
paraview/plugins dir).
This mechanism is necessary to properly install LidarSlamPlugin in
LidarView on Windows so that plugin is auto-loaded and LidarSlam.dll is
found.