[fix] Roll map to center of new frame bounds
Before this commit, if the new frame does not fit at all in all directions (if min point and max point are out of the grid),
the grid would firstly roll to include the min point, then roll to include the max point.
This behavior resulted in always discarding the min point to the benefit of the max one.

Now, if the new frame does not fit into grid, the grid is rolled by the mean of the offsets.

This allows to properly run SLAM when using smaller grids (e.g. indoor).