[fix] Check that time scale is roughly increasing to perform frame undistortion using previous motion
If points timestamps were not excatly computed, it could happen that FrameDuration was bigger than NewFrameTime - PreviousFrameTime.
In that case, approximated undistortion using previous motion was not performed as it detected that time wasn't strictly increasing.

Now, we only check that time is roughly increasing and that FrameDuration is roughly the same as the current and previous timestamps difference.