[fix] Ignore 'time' field if frame duration is null
If input cloud does not have 'time' field or if it is empty, the resulting PCL cloud won't have usable point-wise time info.
Without this point-wise timestamp, undistortion is not possible.
If we detect that frame duration is null (meaning that time field is constant), time values are set to 1.