[fix] Change CeresCostFunctions namespace to use different symbols than LV's
Strangely, when the SLAM is compiled within LV, some of the cost functions used were LV's, and not SLAM's.
Since !66, the cost functions are not compatible with LV's, so in these cases, SLAM completely failed.

To ensure that the correct residuals implementations are used, the namespace have been changed so that the symbols differ.
This also add the missing include into the CMakeList to properly export headers.