[feat][fix] Specify custom time range for LinearTransformInterpolator
Allow to specify custom timestamps for H0 and H1 transforms instead of 0 and 1.

Bug fix: time must be passed as double and not as templated type.
Otherwise, if template is float, we might not be able to precisely represent UNIX timestamps.

Bug fix: add checks to avoid numerical issues (if times or transforms are equal)