[fix][perf] Check output points validity only if some input points are invalid
PolyDataToPointCloud() now returns a boolean to know if some input points are invalid.
If all points are valid, we don't need to run some checks when filling outputs.
If some points are invalid, we loop over each point to update it differently wether it is valid or not.
For advanced debug arrays, we set array value to 0 if the considered point is invalid.