[fix] Weight each residual only once
Each residual was weighted several times :
- In the ceres cost function itself, the squared residual was multiplied by weight.
  This resulted in the residual weighted by sqrt(weight).
- In the scaled loss function, the robustified loss was scaled by weight.

For small residuals, this leads to a scaling of ~weight^1.5.
For bigger residuals, this leads to strange behavior of robustifier, as if the scale was also weighted.

Ceres doc suggests to simply use the ScaledLoss function for such behavior.
So this commit removes the weighting inside the ceres residual computation.