[feat] Change invalidation of points
* Before searching for keypoints, an invalidation step is run. This step
ensures the first and last points of scan rays are not taken into
account (because the spherical aspect of Lidar is not considered),
neither the points which correspond to the boundary of an occluded area
or which lay on a very oblique wall relatively to the laser beam.

* By now, this step is performed with non-intuitive coefficients to set
thresholds. With this commit, the unique threshold is the maximum
orientation of such oblique planes relatively to laser beam to consider
a depth gap.