[fix] Use specified points in CheckPCA, not global ones
When checking if neighborhood was approximately flat, we were not comparing the neigborhood points,
but always the first points from the entire cloud !
This lead to completely wrong behavior, selecting only keypoints along the same direction.

Now, we estimate the approximate line linking the first and last points form the neighborhood,
and compare the angle of each local vector to this line.