[feat] Split Matching and optimization
* Split KeypointsRegistration file into LocalOptimizer and 
KeypointsMatcher files.

* LocalOptimizer

   * It is a member of the SLAM class.
   * A residual structure is created in this file. It contains pointers 
to a cost function and a loss function (optimization estimator to remove 
outliers).
   * It receives the ceres LiDAR residuals.
   * It will be able to receive other sensors' residuals.
   * It is entirely cleared at each registration step.
   * It contains a ceres problem pointer (the destructor is updated). 
This problem is created and filled after each new matching step of ICP. 
The problem is conserved in order to be able to recover covariance from 
the optimization after convergence for last ICP iteration.

* KeypointsMatcher

   * It is an object created locally in Ego-Motion and/or Localization 
functions.
   * It builds the LiDAR residuals (as described above) from matching 
between frame keypoints and map neighborhoods.
   * It contains a vector of residuals (directly added in the matching 
result structure).
   * A residual is also added to the matching info structure.