[feat] Add only keyframes points to maps
A frame is promoted to a keyframe if the motion (distance or angle)
since last keyframe is greater than a user-defined threshold.

For the 10 first keyframes, the distance and angle thresholds are linearly
lowered to quickly build an initial reliable map.

Only keyframes points are added to the keypoints maps.
This help reducing the moving objects artefacts,
especially when the lidar is slowly or not moving.