[fix] Remove BUILD_SHARED_LIBS from cache entries
This variable does not need to be a cache entry.
This variable existence is checked in normal variables, then in cache entries.
This allows to easily set this as a normal variable in a parent scope,
without needing to set it as cache entry, which would override all next calls to this variable.

When needed, this variable may be set locally to the desired value.