[fix][ROS] Specify correct build directory for LidarSlam lib
When ROS wrapping is built, catkin scans all directories,
and builds only those detected as ROS packages (with a CMakeLists.txt and a package.xml).

Therefore, the ROS wrapping build entry point is slam/ros_wrapping/lidar_slam/CMakeLists.txt.
But at the node defined in this CMakeLists needs LidarSlam lib, the compilation of this lib must be triggered.
To build LidarSlam lib, the entry point is slam/CMakeLists.txt.
Therefore, in ROS wrapping, we need to do an out-of-source add_subdirectory
by specifying an additional build directory.
This build directory may depend on ROS build system being used.

The previous directory did not work with `catkin build` from catkin-tools package.
Now, it is properly defined to be more robust and work with both `catkin_make` and `catkin build`.