[ci][fix] Split CMAKE_RUNNER_ARGS into specific options
Strangely, $CMAKE_RUNNER_ARGS is badly dereferenced on windows powershell.
If this variable contains more than one argument, it behaves as if all
the args were not given to cmake.
Therefore, all the deps dirs are missing, and cmake configuration fails.
After many hours of investigation, we have not found why powershell/cmake
fails to correctly deal with this command.

A solution described in the following thread solves this, but is windows specific.
https://stackoverflow.com/questions/33994462/powershell-not-passing-arguments-to-cmake-properly

To keep easier and more readable CI script, the CMAKE_RUNNER_ARGS argument
is split into independent optional options that can be specified on the runner.
There are more args, but this easier to maintain and track errors.