[perf] Add NbPoints attribute to RollingGrid to accelerate Size method
The Size method introduced in !124 loops over all voxels of the rolling grid to get its size.
If the GridSize is significant (>~50), looping over all voxels can be quite slow (tens of ms).
To avoid this loop, NbPoints attribute is added and kept up-to-date.
This attribute can be directly used to know the number of points stored in the rolling grid.