[fix] Check that enough neighbors are requested to build the target model
To build a model of the neighborhood, a minimum number of points is needed:
- 2 for a line model
- 3 for a plane model
- 4 for an ellipsoid model

A check is added to verify that enough neighbors are requested to build the target model.
Otherwise, a MatchStatus::BAD_MODEL_PARAMETRIZATION error code is returned.

This also prevents the matching procedure from crashing if no neighbors are requested.