[feat] Include the landmark detector calibration into the SLAM lib
* This calibration was previously done in the wrapping. It was forcing
the covariance rotation in the wrapping. Now the rotation is managed by
G2O internally.

* It was ok for the landmark detections to be taken in base link but it
is less logical for a GPS measurement or an IMU measurement.

* Moreover, as the landmarks can be used only for local optimization,
the covariance rotation is not always needed.

* Calibration is a new member of the landmarks manager. All landmarks
manager have the same calibration transform.

* Now, to use various landmark detectors, one would need to transform
them in the frame of a reference one because only one calibration is
used.

* An enum is added to refer to indices in the graph calibration
parameters