[fix] Correct maps update/parameterization
* The conversions 1D index / 3D indices were bugged because it is 
different for the inner and for the outer voxel grids

* The leaf size, the outer voxel grid resolution and the number of inner 
voxels are linked so the setters are modified to link the values. A 
VoxelWidth value is added. This allows to store the user supplied 
parameter in VoxelResolution and to compute the closest leafsize 
multiple that will actually be used in VoxelWidth