[feat] Add radius to define neighborhood
* To define a keypoint, the points neighborhoods are studied. Those 
neighborhoods were defined only by a number of points. Now, those 
neighborhoods are also defined by a radius. Points are added while the 
radius AND the number of points are not reached.

* Lines were defined on this neighborhood to compute a curvature. This 
change allows to remove the rejection of too small lines because the 
neighborhood must be large enough now.

* This fixes some specific use cases when the Lidar is very 
perpendicular and quite close to a wall, the neighborhoods were small 
(in terms of space) and always rejected.