[fix] Improve GetTworld when pose has data
Before, the external pose at that time was directly outputed, this is the
correct behavior for an IMU measurement as a graph is built between IMU and
SLAM but not for an external pose.
Now, the external pose measurement is only used to extrapolate since the
last SLAM pose before the required time.