[feat][ROS] Add time offset parameter to external sensors
* time_ext_sensor -> time_ext_sensor + offset

* For example, this could be used for the camera because images can be preprocessed and there can be a delay between pixel acquisition and the
actual image message sending time.
So one may want to put a negative value to fill this time gap.

* This is actually used to do a trick for the camera data : camera data can be received after their corresponding timestamp and too late to be
processed in the Lidar SLAM iteration. So adding a positive offset allows to use the previous camera frame as the synchronized frame to
perform the tests until a better solution is found