[feat] Add safe advance iterator method in Utilities
The standard function std::advance has undefined behavior if the iterator is
out of the list range.
So a SafeAdvance method is added in Utilities which checks a stop condition.
Usually, it is used to stop incrementing (or decrementing) the iterator when
it arrives at the end of the list (or the beginning of the list).