[feat] Improve for optional dependencies in Config file of slam
find_dependency() wraps find_dependency() call which is designed to be used in
a Config.cmake file. It forwards the parameters for EXACT, QUIET and REQUIRED
which were passed to the original find_package call.

The issue is that the QUIET option disables messages only for find_package()
if the package cannot be found.
find_dependency() returns always messages if NOT <package>_FOUND.
So we check manually <package>_FOUND in slam Config.cmake file for optional dependencies.